assignment2-systems/requirements.txt
# PyTorch nightly dependencies - 保持精确版本
--extra-index-url https://download.pytorch.org/whl/nightly/cu128
torch==2.10.0.dev20251013+cu128
pytorch-triton==3.5.0+git7416ffcb

# Core dependencies - 使用范围版本
annotated-types>=0.7.0
certifi>=2024.0.0
charset-normalizer>=3.0.0
click>=8.0.0
cloudpickle>=3.0.0
contourpy>=1.0.0
cycler>=0.12.0
einops>=0.8.0
einx>=0.3.0
filelock>=3.0.0
fonttools>=4.50.0
frozendict>=2.4.0
fsspec>=2024.0.0
gitdb>=4.0.0
GitPython>=3.1.0
humanfriendly>=10.0
idna>=3.0
iniconfig>=2.0.0
jaxtyping>=0.3.0
Jinja2>=3.1.0
kiwisolver>=1.4.0
MarkupSafe>=3.0.0
matplotlib>=3.9.0
mpmath>=1.3.0
networkx>=3.0
numpy>=2.0.0
nvidia-cublas-cu12>=12.8.0
nvidia-cuda-cupti-cu12>=12.8.0
nvidia-cuda-nvrtc-cu12>=12.8.0
nvidia-cuda-runtime-cu12>=12.8.0
nvidia-cudnn-cu12>=9.10.0
nvidia-cufft-cu12>=11.3.0
nvidia-cufile-cu12>=1.13.0
nvidia-curand-cu12>=10.3.0
nvidia-cusolver-cu12>=11.7.0
nvidia-cusparse-cu12>=12.5.0
nvidia-cusparselt-cu12>=0.7.0
nvidia-nccl-cu12>=2.27.0
nvidia-nvjitlink-cu12>=12.8.0
nvidia-nvshmem-cu12>=3.3.0
nvidia-nvtx-cu12>=12.8.0
packaging>=24.0
pandas>=2.0.0
pillow>=11.0.0
pluggy>=1.0.0
protobuf>=6.0.0
psutil>=7.0.0
pydantic>=2.0.0
pydantic-core>=2.0.0
Pygments>=2.0.0
pyparsing>=3.0.0
pytest>=8.0.0
python-dateutil>=2.9.0
pytz>=2024.0
PyYAML>=6.0.0
regex>=2024.11.0
requests>=2.32.0
sentry-sdk>=2.0.0
setuptools>=75.0.0
six>=1.16.0
smmap>=5.0.0
submitit>=1.5.0
sympy>=1.13.0
tabulate>=0.9.0
tiktoken>=0.9.0
tqdm>=4.67.0
typing-extensions>=4.10.0
typing-inspection>=0.4.0
tzdata>=2024.0
urllib3>=2.0.0
wadler-lindig>=0.1.0
wandb>=0.19.0

# Local package (editable install)
-e ./cs336-basics
